refactor(tests): Switch to node-uuid to be consistent with other repos
Since we had trouble with the uuid@2 package (the uuid.v4() subtly changed
from returning a binary to returning an array) we decided to go back to
node-uuid which we use in other repos. Even this has changed somewhat
so that we just need a slight helper to generate a uuid in a Buffer
a little more easily.