refactor(tests): Generate test UIDs a different way
Since we have weird dependency issues related to how the reverse backends work
it may be better off to remove uuid/node-uuid completely so we don't have to
solve it in a weird way. This actually ends up slightly cleaner too, though
we may still have to solve this issue at some point in the future. See
https://github.com/mozilla/fxa-auth-db-mem/issues/23#issuecomment-93856934
for further details.